requirements.txt
flask
flask-cors
flask-session
Flask-SQLAlchemy
Werkzeug
sqlalchemy-access
simplejson
python-jose
requests
